+++ packages/core/src/data-editor/stories/row-groups.stories.tsx
@@ -5,10 +5,13 @@ import {
   useGroupMockDataGenerator,
   useMockDataGenerator,
 } from './utils';
-import React from 'react';
+import React, { useState } from 'react';
 import { DataEditor, DataEditorProps } from '../data-editor';
 import { SimpleThemeWrapper } from '../../stories/story-utils';
 import { clearCell } from './data-editor-beautiful.stories';
+import type { GridSelection, GroupContentRow } from '../../data-grid/data-grid-types';
+import { deleteGroupByGroupRowId, findGroupById } from '../../common/groupUtils';
+import { range } from 'lodash';
 
 const defaultProps: Partial<DataEditorProps> = {
   smoothScrollX: true,
@@ -36,12 +39,22 @@ interface RowGroupsProps {
 }
 
 export const RowGroups: React.VFC<RowGroupsProps> = ({ enableGroups }) => {
-  const { setCellValue, getCellContent, cols, setCellValueRaw } = useMockDataGenerator(
-    20,
-    false,
-    false
+  const [rowsCount, setRowsCount] = useState(100);
+  const { setCellValue, cols, setCellValueRaw, colsMap } = useMockDataGenerator(20, false, false);
+  const { groups, toggleGroup, setGroups } = useGroupMockDataGenerator(20, 2);
+
+  const [rowData, setRowData] = React.useState(() => {
+    return range(0, 300).map((_, rowIndex) =>
+      colsMap.map((columnItem, columnItemIndex) => columnItem.getContent(columnItemIndex, rowIndex))
+    );
+  });
+
+  const getCellContent = React.useCallback(
+    ([col, row]) => {
+      return rowData[row][col];
+    },
+    [rowData]
   );
-  const { groups, toggleGroup, setGroups } = useGroupMockDataGenerator(10, 5);
 
   const colsWithWidths = React.useMemo(() => {
     const c = [...cols];
@@ -52,12 +65,27 @@ export const RowGroups: React.VFC<RowGroupsProps> = ({ enableGroups }) => {
     return c;
   }, [cols]);
 
-  const mangledGetCellContent = React.useCallback(
-    (cellLocation) => {
-      return getCellContent(cellLocation);
+  const onDelete = React.useCallback(
+    (rowIds: GridSelection, deleteGroups: GroupContentRow[]) => {
+      const newGroups = structuredClone(groups);
+      deleteGroups.forEach((item) => {
+        const group = findGroupById(newGroups, item.groupId);
+        if (group !== undefined) {
+          group.rowsCount--;
+          if (group.rowsCount === 0) {
+            deleteGroupByGroupRowId(newGroups, group.id);
+          }
+        }
+      });
+
+      setRowData((oldRowData) => oldRowData.filter((item, index) => !rowIds.rows.hasIndex(index)));
+      setRowsCount((oldRowsCount) => oldRowsCount - rowIds.rows.length);
+      setGroups(newGroups);
+      return true;
     },
-    [getCellContent]
+    [groups, setGroups]
   );
+
   const mangledEditCell = React.useCallback(
     (cellLocation, cell) => setCellValue(cellLocation, cell),
     [setCellValue]
@@ -92,9 +120,9 @@ export const RowGroups: React.VFC<RowGroupsProps> = ({ enableGroups }) => {
     >
       <DataEditor
         {...defaultProps}
-        getCellContent={mangledGetCellContent}
+        getCellContent={getCellContent}
         columns={colsWithWidths}
-        rows={100}
+        rows={rowsCount}
         verticalBorder={false}
         rowMarkers="both"
         onCellEdited={mangledEditCell}
@@ -102,6 +130,7 @@ export const RowGroups: React.VFC<RowGroupsProps> = ({ enableGroups }) => {
         onGroupToggle={toggleGroup}
         freezeColumns={1}
         onRowAppended={onRowAppended}
+        onDelete={onDelete}
         rowMarkerWidth={45}
         trailingRowOptions={{
           sticky: true,